Add restart policy for rsync container
An issue appears when running the e2e CDI tests with podman[1].
The rsync container hits an error as the rsync daemon
fails to attach to the specified port. Adding a restart policy to
restart the container any time it stops allows the e2e lanes to run
successfully using podman.

Signed-off-by: Brian Carey <bcarey@redhat.com>