Wait for podman stop to successfully stop rsync container
bazel-docker.sh is run numerous times during the large e2e jobs. Each
time the script is run a temporary rsync container is started. If one of
these rsync containers is not cleaned up correctly it can lead to rsync
daemon failuresi[1] when e2e jobs are run with podman.

podman stop[2] attempts to stop a container using SIGTERM if this fails
within the default 10 seconds a SIGKILL is used to forceably stop the
container. Previously the script was not waiting for the container to be
successfully stopped.

Also updating the wait before the SIGKILL to 1 second.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/pr-logs/pull/kubevirt_containerized-data-importer/2446/pull-containerized-data-importer-e2e-destructive/1582345804016783360#1:build-log.txt%3A1951
[2] https://docs.podman.io/en/latest/markdown/podman-stop.1.html#description

Signed-off-by: Brian Carey <bcarey@redhat.com>