Suppress alerting for unsupported provisioners (#2476)
Certain storage provisioners are known not to work with CDI.  The main
example is provisioners that implement object storage (ie. buckets).  We
don't want to populate the StorageProfile with ClaimPropertySets but
there is also no need to raise an "incomplete" alert about such a
StorageProfile.  The simplest approach is to simply not count these
unsupported provisioners when setting the IncompleteProfileGauge.

Signed-off-by: Adam Litke <alitke@redhat.com>

Signed-off-by: Adam Litke <alitke@redhat.com>