Update to include mount for containers dir in bazel-docker.sh (#2479)
The quay registry credentials are no longer available when trying to
push images using the podman based bootstrap[1]. This is due to the
bootstrap image now symlinking .docker/config.json to
containers/auth.json [2]. bazel-push expects the registry credentials to
exist under .docker/config.json but the symlink is broken as the
containers directory is not also mounted into the builder container.

This update mounts the container directory into the builder container so
that the symlink can work again.

[1] https://prow.ci.kubevirt.io/view/gs/kubevirt-prow/logs/periodic-containerized-data-importer-push-nightly-ARM64/1594685597958541312
[2] https://github.com/kubevirt/project-infra/pull/2319

Signed-off-by: Brian Carey <bcarey@redhat.com>

Signed-off-by: Brian Carey <bcarey@redhat.com>