Make cluster-sync more friendly for external provider (#2480)
The aim of this change is to make setting up a dev environment on external cluster easier,
By doing mainly 2 things:
- Bazel push (CDI containers/e2e testing containers) expects external registry creds to be in
the docker CRI default path. This seems to be half-configurable on the bazel side,
except for the actual file name (which differs for podman).
Now we prompt for the credentials and save them on the expected path.
- Deploy the e2e testing manifests on external provider as well
Not sure why not do that today? probably reason dates way back

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>