[release-v1.55] Ensure Prometheus resources exist for CDINotReady
Manual backport of #2546 and #2557

Ensure Prometheus resources exist also when cdi-deployment is not
ready. This is needed because currently when cdi deployment fails
(e.g. wrong NodeSelector) the Prometheus resources are not created,
so CDINotReady will not be fired although CDI is not ready.

Also fix StorageProfile incomplete metric test flakiness.

Signed-off-by: Arnon Gilboa <agilboa@redhat.com>