Remove v1alpha1 version from CDI CRD (#2527)
* Tolerate errors due to deployment resourceVersion changing

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Watch and remove v1alpha1 version from CDI CRD

Previously we didn't monitor this resource at all, and since it's
not controlled by the cdi operator, we need to use custom watch
code for it.

Re-use the code for removing old version to make sure that if
v1alpha1 was ever a storedVersion, it's removed. Add a test for
that, too.

Signed-off-by: Maya Rashish <mrashish@redhat.com>

* Don't wait for the deployment to be ready

Signed-off-by: Maya Rashish <mrashish@redhat.com>

---------

Signed-off-by: Maya Rashish <mrashish@redhat.com>