Stop forcing preallocation on block imports (#2588)
Due to what seemed like a bug in qemu-img, we started forcing preallocation on block imports as an arbitrary way to skip the error.

After several updates in qemu and related dependencies, the bug is no longer reproducible.

This commit reverts this behavior and enables deactivating preallocation on block imports.

Signed-off-by: Alvaro Romero <alromero@redhat.com>