Get rid of sync/updateStatus shared state (#2590)
There should be no state shared between sync() and updateStatus().
updateStatus() should stand on it's own, and come to it's own
conclusions based on what it observes. It is okay if it is "late to the
party" and does not observe the latest changes in sync(). It will
eventually converge. This is what kubevirt does.

Signed-off-by: Arnon Gilboa <agilboa@redhat.com>