[release-v1.55] Handle 0 restore size better v155 (#2727)
* Allow restoreSize to be 0, and use request size if 0 (#2680)

Some CSI provisioner when making volume snapshots don't set the
restore size because it is hard to figure out the value or the
actual stored size on the storage is 0 (Copy on Write). There was
a strict check in place in case of CSI drivers not allowing a
restore to a size other than the restoreSize of the snapshot.

This makes the check more lenient that when we encounter a 0 size
restoreSize, we can assume we can use the requested size value in
the pvc

Signed-off-by: Alexander Wels <awels@redhat.com>

* Adjust for 1.55 release

Signed-off-by: Alexander Wels <awels@redhat.com>

---------

Signed-off-by: Alexander Wels <awels@redhat.com>