[release-v1.56] Fix DataImportCron PVC update error check
Fixed in #2696 in main/release-v1.57 as part of the linter error fixes.

PVC lastUseTime annotation was sometimes not updated so DataImportCron
PVC garbage collection was misbehaving and deleting the latest imported
PVC, which was then re-imported again until the update succeeded.

The issue was introduced by #2539 and found due to flaky [test_id:7406].

bz #2221913

Signed-off-by: Arnon Gilboa <agilboa@redhat.com>