Always use scratchspace when importing and converting (#2832) (#2841)
* Remove older nbdkit



* When converting always use scratch space importing instead
of ndbkit. Once we are able to get nbdkit 1.35.8 or newer
we can revert this change since that will include improvements
to the downloading speed.



* Disable metrics test for import because straight import doesn't
return total, and this means the metrics are disabled.



* Fix broken functional tests



* Address review comments



* Additional review comments.
Fixed functional test that was not doing the right
thing while running the test.



* Always set preallocation on block devices when directly
writing to the device



---------

Signed-off-by: Alexander Wels <awels@redhat.com>