[release-v1.56] Do not fail DV reconcile if storage class is not found
Manual backport of #2860, fixing bz #2232347

Currently there is a repeating reconcile error due to exponential
backoff: "DataVolume.storage spec is missing accessMode and no
storageClass to choose profile", however the error is not needed as we
already have SC watch for this case, e.g. when the DV is using the
default SC but no default SC is configured yet.

Note the SC Watch() was mistakenly looking for DVs with unset phase
(like in the main/1.57 branch), but should look for Pending DVs instead.

Signed-off-by: Arnon Gilboa <agilboa@redhat.com>