Address error condition when scratch space is needed to avoid VM failures
When scratch space is needed during imports/uploads, the pod typically fails so it can be restarted with the required space. This is expected behavior that's automatically handled by CDI.

However, since we use an error exit code, this behavior leads to unwanted VM statuses during provisioning in Kubevirt.

This commit addresses this behavior by using a different DV condition, so kubevirt ignores it and the VM can be provisioned as expected.

Signed-off-by: Alvaro Romero <alromero@redhat.com>