Check PVC Prime before updating DV status to improve status reporting (#2928)
With the addition of populators into CDI, we introduced a minor divergence in the DV status reporting between the new and the old flow to avoid checking PVC Prime. Though it would be ideal to completely ignore the PVC Prime in the DV layer, this divergence might confuse some users.

This commit updates the status reporting mechanism so both the old flow and the populator flow match.

Signed-off-by: Alvaro Romero <alromero@redhat.com>