[release-v1.57] Fix DataImportCron PVC GC race and test flakiness (#3064)
Manual backport of #3057

PVC should be timestamped in creation and not only upon import
completion, as it might be mistakenly GCed. LRU sort will choose
PVC with empty timestamp as the first candidate for deletion.
The PVC will be recreated by the controller and eventually
timestamped, so this bug was hidden for a while.

Fixes CNV-37181

Signed-off-by: Arnon Gilboa <agilboa@redhat.com>