Create rpm/README.md (#3090)
* Create README.md

issue #3089 It would be helpful for there to be a README.md in the rpm directory instructing maintainers on the configuration of bazeldnf parameters to populate the rpm lists in rpm/BUILD.bazel via the files .bazelrc, repo.yaml and hack/build/rpm-defs.sh, the fact that bazeldnf needs to be invoked through the make target rpm-defs, and also an indication that bazeldnf needs to be installed on the build host in order as it is not present in the builder image.

Signed-off-by: cfillekes <cfilleke@redhat.com>

* Update rpm/README.md

Correct description of how bazeldnf is installed and invoked in the course of make rpm-defs

Signed-off-by: cfillekes <cfilleke@redhat.com>

* Update rpm/README.md

document use of  ./hack/build/bazel-docker.sh bash to dig around in the builder container. 

Signed-off-by: cfillekes <cfilleke@redhat.com>

---------

Signed-off-by: cfillekes <cfilleke@redhat.com>