Improve handling of bound temp PVC in clone populator (#3126)
There is a short window of time where a bound PVC can have a unpopulated "capacity" field in its status.

The clone-populator handles this case by returning an error. This approach is overkill as the issue is usually fixed instantly.

This commit improves the handling of this case and adds a unit tests.

Signed-off-by: Alvaro Romero <alromero@redhat.com>