Fix volume/access mode inferring for temp host assisted source PVC in snapshot clones (#3155)
* Clone from snapshot: fix volume/access mode inferring for temp host assisted source PVC

Sometimes with snapshot cloning we have to fall back to host assisted.
To do this, we create a temporary restore from the snapshot and initiate a host assisted clone
from that -> target PVC.
The issue this commit fixes is that we set the wrong access/volume modes on this temporary restore PVC
(we set it to the target's).

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>

* Annotate dataimportcron-created snapshot with their source volume mode

This would then get capitalized on by host assisted fallbacks from snapshot cloning

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>

---------

Signed-off-by: Alex Kalenyuk <akalenyu@redhat.com>