Stop handling error phase as "unknown" in clone-populator (#3163)
Clone populator handles its error phase as unknown when it comes to updating the DataVolume. This has some benefits since we avoid setting the DV to failed, which break consistency with the non-populator flow. However, handling it as unknown is a bit vague and we would benefit from actually triggering a DataVolume event if necessary.

This commit updates this behavior so we handle the error accordingly.

Signed-off-by: Alvaro Romero <alromero@redhat.com>